Remove X-Frame-Options for app receipt on dnt.no
Haproxy can not add a response header based on the value of a request
header due to the fact that the request headers are not avaialbe when
`rspadd` (response add header directive) is evaluated.

The way we work arround this limitation is to conditionaly redirecting
to a backend which does not adde the `X-Frame-Options`. Other traffic
will continue to the default backend as normal.

http://www.serverphorums.com/read.php?10,1081270